Fixed russian localizarion.
Removed dots from abbreviated month names. They do not belong here. If and when they are needed they should be a part of format string and not a part of month name. 

Fixed grammatical case for may. It should be nominative here (май) and not genitive (мая).

Shortened abbreviated month names to 3 letters. This is much more common